refactor: split usecases into packages
