+++ app/src/main/java/com/amsterdam/aflami/di/useCaceModule.kt
@@ -1,22 +1,22 @@
 package com.amsterdam.aflami.di
 
-import com.amsterdam.domain.useCase.AddWatchHistoryUseCase
-import com.amsterdam.domain.useCase.GetAndFilterMoviesByKeywordUseCase
-import com.amsterdam.domain.useCase.GetAndFilterTvShowsByKeywordUseCase
-import com.amsterdam.domain.useCase.GetContinueWatchingMoviesUseCase
-import com.amsterdam.domain.useCase.GetEpisodesBySeasonNumberUseCase
-import com.amsterdam.domain.useCase.GetHomeScreenDataUseCase
-import com.amsterdam.domain.useCase.GetMovieCastUseCase
-import com.amsterdam.domain.useCase.GetMovieDetailsUseCase
-import com.amsterdam.domain.useCase.GetMoviesByActorUseCase
-import com.amsterdam.domain.useCase.GetMoviesByCountryUseCase
-import com.amsterdam.domain.useCase.GetMoviesByMoodUseCase
-import com.amsterdam.domain.useCase.GetPopularMoviesUseCase
-import com.amsterdam.domain.useCase.GetSuggestedCountriesUseCase
-import com.amsterdam.domain.useCase.GetTopRatedMoviesUseCase
-import com.amsterdam.domain.useCase.GetTvShowDetailsUseCase
-import com.amsterdam.domain.useCase.GetUpcomingMoviesUseCase
-import com.amsterdam.domain.useCase.RecentSearchesUseCase
+import com.amsterdam.domain.useCase.common.AddWatchHistoryUseCase
+import com.amsterdam.domain.useCase.search.GetAndFilterMoviesByKeywordUseCase
+import com.amsterdam.domain.useCase.search.GetAndFilterTvShowsByKeywordUseCase
+import com.amsterdam.domain.useCase.home.GetContinueWatchingMoviesUseCase
+import com.amsterdam.domain.useCase.details.GetEpisodesBySeasonNumberUseCase
+import com.amsterdam.domain.useCase.home.GetHomeScreenDataUseCase
+import com.amsterdam.domain.useCase.common.GetMovieCastUseCase
+import com.amsterdam.domain.useCase.details.GetMovieDetailsUseCase
+import com.amsterdam.domain.useCase.search.GetMoviesByActorUseCase
+import com.amsterdam.domain.useCase.search.GetMoviesByCountryUseCase
+import com.amsterdam.domain.useCase.home.GetMoviesByMoodUseCase
+import com.amsterdam.domain.useCase.home.GetPopularMoviesUseCase
+import com.amsterdam.domain.useCase.search.GetSuggestedCountriesUseCase
+import com.amsterdam.domain.useCase.home.GetTopRatedMoviesUseCase
+import com.amsterdam.domain.useCase.details.GetTvShowDetailsUseCase
+import com.amsterdam.domain.useCase.home.GetUpcomingMoviesUseCase
+import com.amsterdam.domain.useCase.search.RecentSearchesUseCase
 import com.amsterdam.domain.useCase.authentication.GetsSessionType
 import com.amsterdam.domain.useCase.authentication.LoginAsGuestUseCase
 import com.amsterdam.domain.useCase.authentication.LoginWithPasswordUseCase
