+++ domain/src/main/java/com/amsterdam/domain/useCase/common/GetMovieCastUseCase.kt
@@ -1,4 +1,4 @@
-package com.amsterdam.domain.useCase
+package com.amsterdam.domain.useCase.common
 
 import com.amsterdam.domain.repository.MovieRepository
 import com.amsterdam.entity.Actor
