+++ domain/src/main/java/com/amsterdam/domain/useCase/details/GetEpisodesBySeasonNumberUseCase.kt
@@ -1,4 +1,4 @@
-package com.amsterdam.domain.useCase
+package com.amsterdam.domain.useCase.details
 
 import com.amsterdam.domain.repository.TvShowRepository
 import com.amsterdam.entity.Episode
