+++ domain/src/main/java/com/amsterdam/domain/useCase/home/GetContinueWatchingMoviesUseCase.kt
@@ -1,4 +1,4 @@
-package com.amsterdam.domain.useCase
+package com.amsterdam.domain.useCase.home
 
 import com.amsterdam.domain.repository.WatchHistoryRepository
 import com.amsterdam.entity.Movie
