+++ domain/src/main/java/com/amsterdam/domain/useCase/home/GetHomeScreenDataUseCase.kt
@@ -1,4 +1,4 @@
-package com.amsterdam.domain.useCase
+package com.amsterdam.domain.useCase.home
 
 import com.amsterdam.entity.Movie
 import com.amsterdam.entity.category.MovieGenre
