+++ domain/src/main/java/com/amsterdam/domain/useCase/home/GetUpcomingMoviesUseCase.kt
@@ -1,4 +1,4 @@
-package com.amsterdam.domain.useCase
+package com.amsterdam.domain.useCase.home
 
 import com.amsterdam.domain.repository.MovieRepository
 import com.amsterdam.entity.Movie
