+++ domain/src/main/java/com/amsterdam/domain/useCase/search/RecentSearchesUseCase.kt
@@ -1,4 +1,4 @@
-package com.amsterdam.domain.useCase
+package com.amsterdam.domain.useCase.search
 
 import com.amsterdam.domain.repository.RecentSearchRepository
 import com.amsterdam.entity.Country
