+++ domain/src/test/java/com/amsterdam/domain/useCase/GetMovieDetailsUseCaseTest.kt
@@ -3,6 +3,7 @@ package com.amsterdam.domain.useCase
 
 import com.amsterdam.domain.exceptions.AflamiException
 import com.amsterdam.domain.repository.MovieRepository
+import com.amsterdam.domain.useCase.details.GetMovieDetailsUseCase
 import com.amsterdam.entity.Actor
 import com.amsterdam.entity.Gender
 import com.amsterdam.entity.Movie
