+++ domain/src/test/java/com/amsterdam/domain/useCase/GetMoviesByActorUseCaseTest.kt
@@ -2,6 +2,7 @@ package com.amsterdam.domain.useCase
 
 import com.amsterdam.domain.exceptions.AflamiException
 import com.amsterdam.domain.repository.MovieRepository
+import com.amsterdam.domain.useCase.search.GetMoviesByActorUseCase
 import com.amsterdam.domain.useCase.utils.fakeMovieList
 import com.google.common.truth.Truth.assertThat
 import io.mockk.coEvery
