+++ domain/src/test/java/com/amsterdam/domain/useCase/GetPopularMoviesUseCaseTest.kt
@@ -2,6 +2,7 @@ package com.amsterdam.domain.useCase
 
 import com.amsterdam.domain.exceptions.AflamiException
 import com.amsterdam.domain.repository.MovieRepository
+import com.amsterdam.domain.useCase.home.GetPopularMoviesUseCase
 import com.amsterdam.domain.useCase.utils.fakeMovieList
 import com.google.common.truth.Truth.assertThat
 import io.mockk.coEvery
