+++ domain/src/test/java/com/amsterdam/domain/useCase/GetTopRatedMoviesUseCaseTest.kt
@@ -1,6 +1,7 @@
 package com.amsterdam.domain.useCase
 
 import com.amsterdam.domain.repository.MovieRepository
+import com.amsterdam.domain.useCase.home.GetTopRatedMoviesUseCase
 import com.amsterdam.domain.useCase.utils.specificMovieList
 import com.google.common.truth.Truth.assertThat
 import io.mockk.coEvery
