+++ viewModel/src/test/java/com/amsterdam/viewmodel/SearchByCountryViewModelTest.kt
@@ -3,9 +3,9 @@ package com.amsterdam.viewmodel
 
 import com.amsterdam.domain.exceptions.AflamiException
 import com.amsterdam.domain.exceptions.NoInternetException
-import com.amsterdam.domain.useCase.GetMoviesByCountryUseCase
-import com.amsterdam.domain.useCase.GetSuggestedCountriesUseCase
-import com.amsterdam.domain.useCase.RecentSearchesUseCase
+import com.amsterdam.domain.useCase.search.GetMoviesByCountryUseCase
+import com.amsterdam.domain.useCase.search.GetSuggestedCountriesUseCase
+import com.amsterdam.domain.useCase.search.RecentSearchesUseCase
 import com.amsterdam.entity.Country
 import com.amsterdam.viewmodel.search.countrySearch.CountryItemUiState
 import com.amsterdam.viewmodel.search.countrySearch.CountrySearchEffect
