+++ viewModel/src/test/java/com/amsterdam/viewmodel/movieDetails/MovieDetailsViewModelTest.kt
@@ -2,7 +2,7 @@ package com.amsterdam.viewmodel.movieDetails
 
 import com.amsterdam.domain.exceptions.AflamiException
 import com.amsterdam.domain.exceptions.NoInternetException
-import com.amsterdam.domain.useCase.GetMovieDetailsUseCase
+import com.amsterdam.domain.useCase.details.GetMovieDetailsUseCase
 import com.amsterdam.viewmodel.movieDetails.MovieDetailsUiState.MovieExtras
 import com.amsterdam.viewmodel.shared.Selectable
 import com.amsterdam.viewmodel.utils.TestDispatcherProvider
